Deletes GoNext.
