+++ financial-connections/src/main/java/com/stripe/android/financialconnections/di/FinancialConnectionsSheetNativeModule.kt
@@ -27,6 +27,7 @@ import com.stripe.android.financialconnections.repository.api.FinancialConnectio
 import com.stripe.android.repository.ConsumersApiService
 import com.stripe.android.repository.ConsumersApiServiceImpl
 import com.stripe.android.uicore.image.StripeImageLoader
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import kotlinx.coroutines.CoroutineScope
@@ -48,111 +49,111 @@ import javax.inject.Singleton
         ResetSubcomponent::class
     ]
 )
-internal class FinancialConnectionsSheetNativeModule {
+internal interface FinancialConnectionsSheetNativeModule {
 
     @Singleton
-    @Provides
+    @Binds
     fun providesNavigationManager(
-        logger: Logger
-    ): NavigationManager = NavigationManagerImpl(
-        logger = logger
-    )
+        impl: NavigationManagerImpl
+    ): NavigationManager
 
-    @Provides
-    @Singleton
-    fun provideConsumersApiService(
-        apiVersion: ApiVersion,
-        stripeNetworkClient: StripeNetworkClient,
-    ): ConsumersApiService = ConsumersApiServiceImpl(
-        appInfo = null,
-        sdkVersion = StripeSdkVersion.VERSION,
-        apiVersion = apiVersion.code,
-        stripeNetworkClient = stripeNetworkClient
-    )
+    companion object {
+        @Provides
+        @Singleton
+        fun provideConsumersApiService(
+            apiVersion: ApiVersion,
+            stripeNetworkClient: StripeNetworkClient,
+        ): ConsumersApiService = ConsumersApiServiceImpl(
+            appInfo = null,
+            sdkVersion = StripeSdkVersion.VERSION,
+            apiVersion = apiVersion.code,
+            stripeNetworkClient = stripeNetworkClient
+        )
 
-    @Singleton
-    @Provides
-    fun providesImageLoader(
-        context: Application
-    ) = StripeImageLoader(
-        context = context,
-        diskCache = null,
-    )
+        @Singleton
+        @Provides
+        fun providesImageLoader(
+            context: Application
+        ) = StripeImageLoader(
+            context = context,
+            diskCache = null,
+        )
 
-    @Singleton
-    @Provides
-    fun providesFinancialConnectionsManifestRepository(
-        requestExecutor: FinancialConnectionsRequestExecutor,
-        apiRequestFactory: ApiRequest.Factory,
-        apiOptions: ApiRequest.Options,
-        locale: Locale?,
-        logger: Logger,
-        @Named(INITIAL_SYNC_RESPONSE) initialSynchronizeSessionResponse: SynchronizeSessionResponse?
-    ) = FinancialConnectionsManifestRepository(
-        requestExecutor = requestExecutor,
-        apiRequestFactory = apiRequestFactory,
-        apiOptions = apiOptions,
-        locale = locale ?: Locale.getDefault(),
-        logger = logger,
-        initialSync = initialSynchronizeSessionResponse
-    )
+        @Singleton
+        @Provides
+        fun providesFinancialConnectionsManifestRepository(
+            requestExecutor: FinancialConnectionsRequestExecutor,
+            apiRequestFactory: ApiRequest.Factory,
+            apiOptions: ApiRequest.Options,
+            locale: Locale?,
+            logger: Logger,
+            @Named(INITIAL_SYNC_RESPONSE) initialSynchronizeSessionResponse: SynchronizeSessionResponse?
+        ) = FinancialConnectionsManifestRepository(
+            requestExecutor = requestExecutor,
+            apiRequestFactory = apiRequestFactory,
+            apiOptions = apiOptions,
+            locale = locale ?: Locale.getDefault(),
+            logger = logger,
+            initialSync = initialSynchronizeSessionResponse
+        )
 
-    @Singleton
-    @Provides
-    fun providesFinancialConnectionsConsumerSessionRepository(
-        consumersApiService: ConsumersApiService,
-        apiOptions: ApiRequest.Options,
-        financialConnectionsConsumersApiService: FinancialConnectionsConsumersApiService,
-        locale: Locale?,
-        logger: Logger,
-    ) = FinancialConnectionsConsumerSessionRepository(
-        financialConnectionsConsumersApiService = financialConnectionsConsumersApiService,
-        consumersApiService = consumersApiService,
-        apiOptions = apiOptions,
-        locale = locale ?: Locale.getDefault(),
-        logger = logger,
-    )
+        @Singleton
+        @Provides
+        fun providesFinancialConnectionsConsumerSessionRepository(
+            consumersApiService: ConsumersApiService,
+            apiOptions: ApiRequest.Options,
+            financialConnectionsConsumersApiService: FinancialConnectionsConsumersApiService,
+            locale: Locale?,
+            logger: Logger,
+        ) = FinancialConnectionsConsumerSessionRepository(
+            financialConnectionsConsumersApiService = financialConnectionsConsumersApiService,
+            consumersApiService = consumersApiService,
+            apiOptions = apiOptions,
+            locale = locale ?: Locale.getDefault(),
+            logger = logger,
+        )
 
-    @Singleton
-    @Provides
-    fun providesFinancialConnectionsAccountsRepository(
-        requestExecutor: FinancialConnectionsRequestExecutor,
-        apiOptions: ApiRequest.Options,
-        apiRequestFactory: ApiRequest.Factory,
-        logger: Logger
-    ) = FinancialConnectionsAccountsRepository(
-        requestExecutor = requestExecutor,
-        apiRequestFactory = apiRequestFactory,
-        apiOptions = apiOptions,
-        logger = logger
-    )
+        @Singleton
+        @Provides
+        fun providesFinancialConnectionsAccountsRepository(
+            requestExecutor: FinancialConnectionsRequestExecutor,
+            apiOptions: ApiRequest.Options,
+            apiRequestFactory: ApiRequest.Factory,
+            logger: Logger
+        ) = FinancialConnectionsAccountsRepository(
+            requestExecutor = requestExecutor,
+            apiRequestFactory = apiRequestFactory,
+            apiOptions = apiOptions,
+            logger = logger
+        )
 
-    @Singleton
-    @Provides
-    fun providesFinancialConnectionsInstitutionsRepository(
-        requestExecutor: FinancialConnectionsRequestExecutor,
-        apiRequestFactory: ApiRequest.Factory,
-        apiOptions: ApiRequest.Options
-    ) = FinancialConnectionsInstitutionsRepository(
-        requestExecutor = requestExecutor,
-        apiOptions = apiOptions,
-        apiRequestFactory = apiRequestFactory
-    )
+        @Singleton
+        @Provides
+        fun providesFinancialConnectionsInstitutionsRepository(
+            requestExecutor: FinancialConnectionsRequestExecutor,
+            apiRequestFactory: ApiRequest.Factory,
+            apiOptions: ApiRequest.Options
+        ) = FinancialConnectionsInstitutionsRepository(
+            requestExecutor = requestExecutor,
+            apiOptions = apiOptions,
+            apiRequestFactory = apiRequestFactory
+        )
 
-    @Singleton
-    @Provides
-    fun providesSaveToLinkWithStripeSucceededRepository() = SaveToLinkWithStripeSucceededRepository(
-        CoroutineScope(SupervisorJob() + Dispatchers.Default)
-    )
+        @Singleton
+        @Provides
+        fun providesSaveToLinkWithStripeSucceededRepository() = SaveToLinkWithStripeSucceededRepository(
+            CoroutineScope(SupervisorJob() + Dispatchers.Default)
+        )
 
-    @Provides
-    internal fun provideFinancialConnectionsConsumersApiService(
-        requestExecutor: FinancialConnectionsRequestExecutor,
-        apiOptions: ApiRequest.Options,
-        apiRequestFactory: ApiRequest.Factory,
-    ): FinancialConnectionsConsumersApiService = FinancialConnectionsConsumersApiService(
-        apiOptions = apiOptions,
-        apiRequestFactory = apiRequestFactory,
-        requestExecutor = requestExecutor
-    )
+        @Provides
+        internal fun provideFinancialConnectionsConsumersApiService(
+            requestExecutor: FinancialConnectionsRequestExecutor,
+            apiOptions: ApiRequest.Options,
+            apiRequestFactory: ApiRequest.Factory,
+        ): FinancialConnectionsConsumersApiService = FinancialConnectionsConsumersApiService(
+            apiOptions = apiOptions,
+            apiRequestFactory = apiRequestFactory,
+            requestExecutor = requestExecutor
+        )
+    }
 }
